separating utility function from main code
